Remove most unused code.
